Renamed Feature to Counter
I've also added an App container. Later to be moved to a container
component file structure.
