+++ app/config/routes.js
@@ -1,10 +1,11 @@
 import React from 'react'
 import { Route, IndexRoute } from 'react-router'
-import { Feature, About, MoreInfo, Home } from '../modules'
+import { App, About, MoreInfo, Home } from '../modules'
 
 export default (
-  <Route path="/" component={Feature}>
+  <Route path="/" component={App}>
     <IndexRoute component={Home} />
+
     <Route path="about" component={About} />
     <Route path="more" component={MoreInfo} />
   </Route>
