+++ app/modules/index.js
@@ -1,11 +1,11 @@
-import About from './about/component'
-import Feature from './feature/component'
-import MoreInfo from './more/component'
+import App from './app/component'
 import Home from './home/component'
+import About from './about/component'
+import MoreInfo from './more/component'
 
 export {
+  App,
+  Home,
   About,
-  Feature,
-  MoreInfo,
-  Home
+  MoreInfo
 }
